Fix wrong udev rule (better described at #7)
Starting serialjs with systemd-run ensures it wouldn't be killed by udev
after a short time.
However, this solution makes it impossible to handle multiple gamepads
and boards at the same time since we are killing all serialjs instances
before spawning one. This is a temporary solution and I will fix it
later as described in #7.